Move all files to a module
