+++ tasks/__init__.py
@@ -71,7 +71,7 @@ from tasks.go_test import (
 )
 from tasks.install_tasks import download_tools, install_shellcheck, install_tools
 from tasks.junit_tasks import junit_upload
-from tasks.libs.go_workspaces import handle_go_work
+from tasks.libs.common.go_workspaces import handle_go_work
 from tasks.pr_checks import lint_releasenote
 from tasks.show_linters_issues import show_linters_issues
 from tasks.unit_tests import invoke_unit_tests
