+++ tasks/unit-tests/github_api_tests.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import patch
 
-from tasks.libs.common.github_api import GithubAPI
+from tasks.libs.ciproviders.github_api import GithubAPI
 
 
 class Label:
@@ -15,7 +15,7 @@ class Myfile:
 
 
 class TestReleaseNoteNeeded(unittest.TestCase):
-    @patch("tasks.libs.common.github_api.Github", autospec=True)
+    @patch("tasks.libs.ciproviders.github_api.Github", autospec=True)
     def test_needed(self, _):
         github = GithubAPI(repository="test", public_repo=True)
         with patch.object(github, '_repository') as mock_repo:
@@ -23,7 +23,7 @@ class TestReleaseNoteNeeded(unittest.TestCase):
             mock_pr.get_labels.return_value = [Label("changeling/no-changelog")]
             self.assertTrue(github.is_release_note_needed(1))
 
-    @patch("tasks.libs.common.github_api.Github", autospec=True)
+    @patch("tasks.libs.ciproviders.github_api.Github", autospec=True)
     def test_not_needed(self, _):
         github = GithubAPI(repository="test", public_repo=True)
         with patch.object(github, '_repository') as mock_repo:
@@ -33,7 +33,7 @@ class TestReleaseNoteNeeded(unittest.TestCase):
 
 
 class TestContainsReleaseNote(unittest.TestCase):
-    @patch("tasks.libs.common.github_api.Github", autospec=True)
+    @patch("tasks.libs.ciproviders.github_api.Github", autospec=True)
     def test_contains(self, _):
         github = GithubAPI(repository="test", public_repo=True)
         with patch.object(github, '_repository') as mock_repo:
@@ -41,7 +41,7 @@ class TestContainsReleaseNote(unittest.TestCase):
             mock_pr.get_files.return_value = [Myfile("releasenotes/notes/do/re/mi")]
             self.assertTrue(github.contains_release_note(1))
 
-    @patch("tasks.libs.common.github_api.Github", autospec=True)
+    @patch("tasks.libs.ciproviders.github_api.Github", autospec=True)
     def test_do_not_contain(self, _):
         github = GithubAPI(repository="test", public_repo=True)
         with patch.object(github, '_repository') as mock_repo:
